requirements.txt
pyexiftool
openai>=1.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
dateparser>=1.1.8
pdf2image>=1.16.3